# openapi.yaml
openapi: 3.1.0
info:
  title: LizardNotes API
  version: 1.0.0
  description: Backend API for LizardNotes — a personal Obsidian-style markdown notes app.

# The API URL is environment-specific and is read at deploy time from SSM parameter
# /lizardnotes/apigateway/apiUrl. Replace the placeholder below with the actual URL
# from your deployed infra stack before using this spec with external tooling.
servers:
  - url: https://your-api-id.execute-api.your-region.amazonaws.com
    description: API Gateway HTTP API (value from SSM /lizardnotes/apigateway/apiUrl)
security:
  - CognitoJWT: []
components:
  securitySchemes:
    CognitoJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Attachment:
      type: object
      required:
        - attachmentId
        - noteId
        - filename
        - mimeType
        - size
        - s3Key
        - createdAt
        - userId
      properties:
        attachmentId:
          type: string
        noteId:
          type: string
        filename:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        s3Key:
          type: string
        createdAt:
          type: string
          format: date-time
        userId:
          type: string
    AttachmentWithDownloadUrl:
      allOf:
        - $ref: '#/components/schemas/Attachment'
        - type: object
          required:
            - downloadUrl
          properties:
            downloadUrl:
              type: string
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
  requestBodies:
...
    CreateAttachmentRequest:
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - filename
              - mimeType
              - size
            properties:
              filename:
                type: string
              mimeType:
                type: string
              size:
                type: integer
paths:
...
  /notes/{noteId}/attachments:
    parameters:
      - name: noteId
        in: path
        required: true
        schema:
          type: string

    get:
      operationId: listAttachments
      summary: List attachments for a note, each with a presigned download URL
      responses:
        '200':
          description: List of attachments with presigned download URLs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttachmentWithDownloadUrl'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: createAttachment
      summary: Register an attachment and get a presigned upload URL
      requestBody:
        $ref: '#/components/requestBodies/CreateAttachmentRequest'
      responses:
        '201':
          description: Attachment record and presigned S3 upload URL
          content:
            application/json:
              schema:
                type: object
                required:
                  - attachment
                  - uploadUrl
                properties:
                  attachment:
                    $ref: '#/components/schemas/Attachment'
                  uploadUrl:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /notes/{noteId}/attachments/{attachmentId}:
    parameters:
      - name: noteId
        in: path
        required: true
        schema:
          type: string
      - name: attachmentId
        in: path
        required: true
        schema:
          type: string

    get:
      operationId: getAttachment
      summary: Get attachment metadata and a presigned S3 download URL (60-minute expiry)
      responses:
        '200':
          description: Attachment metadata with presigned download URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttachmentWithDownloadUrl'
        '404':
          description: Attachment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: deleteAttachment
      summary: Delete an attachment record and its S3 object
      responses:
        '204':
          description: Attachment deleted
        '404':
          description: Attachment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'